requirements.txt
astroid~=2.3.3
certifi~=2019.11.28
chardet~=3.0.4
Click~=7.0
gitdb~=4.0.2
GitPython~=3.1.0
idna~=2.9
isort~=4.3.21
Jinja2~=2.11.1
lazy-object-proxy~=1.4.3
MarkupSafe~=1.1.1
mccabe~=0.6.1
passpy~=1.0.2
pylint~=2.4.4
pyperclip~=1.7.0
python-gnupg~=0.4.5
requests~=2.23.0
ruamel.yaml~=0.16.10
ruamel.yaml.clib~=0.2.0
six~=1.14.0
smmap~=3.0.1
urllib3~=1.25.8
wrapt~=1.11.2
